[BUGFIX] Fix typing issue

+++ Classes/Service/StyleguideConfigurationManager.php
@@ -242,7 +242,7 @@ class StyleguideConfigurationManager
                 PathUtility::getAbsoluteWebPath($path, false)
             )
             ->withQuery('?' . $modified)
-            ->withPort(GeneralUtility::getIndpEnv('TYPO3_PORT') ?: null);
+            ->withPort((int) GeneralUtility::getIndpEnv('TYPO3_PORT') ?: null);
     }
 
     /**
